refactor: @Component에서 @TestComponent로 변경
